Centralize query date-time shape handling
